Fix num_workers and ignore some args in cache
num_workers soffices to be just 2, since the forward+backward pass is
relatively slow compared to batch creation, as long as the dataloader is
using multiprocessing (eg num_workers > 0) it should be fine.

For the caching mechanism, since some function received the num_workers
arg, which was set to os.cpu_count(), caches were incompatible between
machines, because a server might have used 8 workers while a dev machine
only 4 or so. Now caches should be compatible with different machines,
and we should be able to rsync caches between machines to compute just
once ever.